Refactor: Split Users page into smaller components
This commit refactors the Users page into smaller, more manageable components to improve maintainability and readability.
